Added abstract class GameObjects. All game objects need to implement getVelocity, and therefore have to have a velocity. Started to implement CircleCollider. Function: overlap - done static with dynamic collision - in progress

+++ 2DGameLib/src/collision/Collider.java
@@ -1,8 +1,8 @@
 package collision;
 
 public abstract class Collider {
-	abstract public int getX();
-	abstract public int getY();
+	abstract public double getX();
+	abstract public double getY();
 	abstract public boolean overlaps(Collider c);
 
 }
